requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.5.0
pathlib==1.0.1
aiofiles==23.2.1
psycopg2-binary==2.9.9